Working now. Transitioned to redux, but some components (Welcome, Question) still need to be connected with react-redux.
